refactor: RacingOptions 객체 제거
