+++ src/main/java/game/service/CarRacing.java
@@ -9,9 +9,17 @@ public class CarRacing {
     private final int racingRep;
     private final Cars cars;
 
-    public CarRacing(RacingOptions options) {
-        this.racingRep = options.racingRep();
-        this.cars = Cars.generate(options.carCount());
+    public CarRacing(int carCount, int racingRep) {
+        throwIfNegativeNumber(carCount);
+        throwIfNegativeNumber(racingRep);
+        this.racingRep = racingRep;
+        this.cars = Cars.generate(carCount);
+    }
+
+    private void throwIfNegativeNumber(int number) {
+        if (number <= 0) {
+            throw new IllegalArgumentException("입력은 0 보다 큰 양수만 가능합니다.");
+        }
     }
 
     public void start() {
